fix: 프로필 토글 버그 수정
